+++ examples/basic/main.tf
@@ -27,10 +27,10 @@ resource "azurerm_virtual_network" "csr" {
 }
 
 resource "azurerm_subnet" "csr" {
+  address_prefixes     = ["10.0.2.0/24"]
   name                 = "sn-private"
   resource_group_name  = azurerm_resource_group.csr.name
   virtual_network_name = azurerm_virtual_network.csr.name
-  address_prefixes     = ["10.0.2.0/24"]
 }
 
 module "csr" {